# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ActionItemOut": {
        "properties": {
          "amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "type": "string"
          },
          "days_overdue": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Overdue"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "kind": {
            "enum": [
              "overdue_invoice",
              "gst_payable",
              "accrual_loss"
            ],
            "title": "Kind",
            "type": "string"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "severity": {
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Severity",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "severity",
          "title",
          "detail",
          "amount",
          "reference",
          "days_overdue",
          "due_date"
        ],
        "title": "ActionItemOut",
        "type": "object"
      },
...
      "BalanceGroup": {
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/BalanceLineOut"
            },
            "title": "Lines",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "total": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total",
            "type": "string"
          }
        },
        "required": [
          "name",
          "total",
          "lines"
        ],
        "title": "BalanceGroup",
        "type": "object"
      },
...
      "BalanceSheet": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "basis": {
            "default": "balance",
            "enum": [
              "accrual",
              "cash",
              "billed",
              "balance",
              "ledger"
            ],
            "title": "Basis",
            "type": "string"
          },
          "cash_on_hand": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cash On Hand",
            "type": "string"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/BalanceGroup"
            },
            "title": "Groups",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "pending_definition": {
            "default": true,
            "description": "Awaiting the director's definition (Q6)",
            "title": "Pending Definition",
            "type": "boolean"
          }
        },
        "required": [
          "meta",
          "as_of",
          "cash_on_hand",
          "groups"
        ],
        "title": "BalanceSheet",
        "type": "object"
      },
...
      "ExpenseShare": {
        "properties": {
          "amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "share": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Percent of the total",
            "title": "Share"
          }
        },
        "required": [
          "name",
          "amount",
          "share"
        ],
        "title": "ExpenseShare",
        "type": "object"
      },
...
      "GstOut": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "basis": {
            "default": "ledger",
            "enum": [
              "accrual",
              "cash",
              "billed",
              "balance",
              "ledger"
            ],
            "title": "Basis",
            "type": "string"
          },
          "input_tax": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tax"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "net_payable": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Payable"
          },
          "output_tax": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tax"
          },
          "pending_definition": {
            "default": true,
            "description": "Awaiting the director's definition (Q5)",
            "title": "Pending Definition",
            "type": "boolean"
          }
        },
        "required": [
          "meta",
          "as_of",
          "output_tax",
          "input_tax",
          "net_payable"
        ],
        "title": "GstOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvoiceOut": {
        "properties": {
          "balance": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Balance",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "days_overdue": {
            "title": "Days Overdue",
            "type": "integer"
          },
          "due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "invoice_date": {
            "format": "date",
            "title": "Invoice Date",
            "type": "string"
          },
          "invoice_number": {
            "title": "Invoice Number",
            "type": "string"
          },
          "party_name": {
            "title": "Party Name",
            "type": "string"
          },
          "status": {
            "enum": [
              "draft",
              "void",
              "paid",
              "partially_paid",
              "open",
              "overdue",
              "unknown"
            ],
            "title": "Status",
            "type": "string"
          },
          "total": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total",
            "type": "string"
          },
          "total_document": {
            "description": "In the invoice's own currency",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Document",
            "type": "string"
          }
        },
        "required": [
          "id",
          "invoice_number",
          "party_name",
          "invoice_date",
          "due_date",
          "total",
          "balance",
          "currency",
          "total_document",
          "status",
          "days_overdue"
        ],
        "title": "InvoiceOut",
        "type": "object"
      },
      "InvoicePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "InvoicePage",
        "type": "object"
      },
      "Kpi": {
        "properties": {
          "basis": {
            "enum": [
              "accrual",
              "cash",
              "billed",
              "balance",
              "ledger"
            ],
            "title": "Basis",
            "type": "string"
          },
          "change": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Change"
              },
              {
                "type": "null"
              }
            ]
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "e.g. overdue invoices for receivables",
            "title": "Count"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "ratio": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Percent: collection ratio, net margin",
            "title": "Ratio"
          },
          "related": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NamedAmount"
              },
              {
                "type": "null"
              }
            ],
            "description": "e.g. {\"billed\": 895675} for collected"
          },
          "value": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Value",
            "type": "string"
          },
          "window": {
            "$ref": "#/components/schemas/Window"
          }
        },
        "required": [
          "key",
          "label",
          "value",
          "basis",
          "window"
        ],
        "title": "Kpi",
        "type": "object"
      },
...
      "Meta": {
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "fiscal_year": {
            "description": "e.g. \"FY 2026-27\"",
            "title": "Fiscal Year",
            "type": "string"
          },
          "fiscal_year_start_month": {
            "title": "Fiscal Year Start Month",
            "type": "integer"
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness"
          },
          "organisation": {
            "title": "Organisation",
            "type": "string"
          },
          "today": {
            "description": "Today in the organisation's time zone",
            "format": "date",
            "title": "Today",
            "type": "string"
          }
        },
        "required": [
          "organisation",
          "currency",
          "fiscal_year",
          "fiscal_year_start_month",
          "today",
          "freshness"
        ],
        "title": "Meta",
        "type": "object"
      },
...
      "Overview": {
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionItemOut"
            },
            "title": "Actions",
            "type": "array"
          },
          "expense_mix": {
            "description": "Operating expenses by account (accrual)",
            "items": {
              "$ref": "#/components/schemas/ExpenseShare"
            },
            "title": "Expense Mix",
            "type": "array"
          },
          "expense_mix_window": {
            "$ref": "#/components/schemas/Window"
          },
          "kpis": {
            "items": {
              "$ref": "#/components/schemas/Kpi"
            },
            "title": "Kpis",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "months": {
            "description": "Billed vs collected vs expenses (cash)",
            "items": {
              "$ref": "#/components/schemas/MonthFigures"
            },
            "title": "Months",
            "type": "array"
          },
          "pnl": {
            "$ref": "#/components/schemas/PnlStatement"
          },
          "register_totals": {
            "$ref": "#/components/schemas/RegisterTotalsOut"
          },
          "revenue_by_client": {
            "items": {
              "$ref": "#/components/schemas/ClientBilled"
            },
            "title": "Revenue By Client",
            "type": "array"
          }
        },
        "required": [
          "meta",
          "kpis",
          "pnl",
          "months",
          "actions",
          "revenue_by_client",
          "expense_mix",
          "expense_mix_window",
          "register_totals"
        ],
        "title": "Overview",
        "type": "object"
      },
...
      "Receivables": {
        "properties": {
          "ageing": {
            "items": {
              "$ref": "#/components/schemas/AgeBucketOut"
            },
            "title": "Ageing",
            "type": "array"
          },
          "basis": {
            "default": "balance",
            "enum": [
              "accrual",
              "cash",
              "billed",
              "balance",
              "ledger"
            ],
            "title": "Basis",
            "type": "string"
          },
          "by_client": {
            "items": {
              "$ref": "#/components/schemas/ClientBilled"
            },
            "title": "By Client",
            "type": "array"
          },
          "invoices": {
            "description": "Every receivable, most overdue first",
            "items": {
              "$ref": "#/components/schemas/InvoiceOut"
            },
            "title": "Invoices",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "open_count": {
            "title": "Open Count",
            "type": "integer"
          },
          "outstanding": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Outstanding",
            "type": "string"
          },
          "overdue": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Overdue",
            "type": "string"
          },
          "overdue_count": {
            "title": "Overdue Count",
            "type": "integer"
          }
        },
        "required": [
          "meta",
          "outstanding",
          "overdue",
          "overdue_count",
          "open_count",
          "ageing",
          "by_client",
          "invoices"
        ],
        "title": "Receivables",
        "type": "object"
      },
...
      "Trends": {
        "properties": {
          "daily": {
            "description": "Last 30 days",
            "items": {
              "$ref": "#/components/schemas/CashBucketOut"
            },
            "title": "Daily",
            "type": "array"
          },
          "expense_trend": {
            "items": {
              "$ref": "#/components/schemas/CategoryMonth"
            },
            "title": "Expense Trend",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "months": {
            "items": {
              "$ref": "#/components/schemas/MonthFigures"
            },
            "title": "Months",
            "type": "array"
          },
          "top_categories": {
            "items": {
              "type": "string"
            },
            "title": "Top Categories",
            "type": "array"
          },
          "totals": {
            "$ref": "#/components/schemas/MonthFigures"
          },
          "weekly": {
            "description": "Last 10 weeks, weeks start on Monday",
            "items": {
              "$ref": "#/components/schemas/CashBucketOut"
            },
            "title": "Weekly",
            "type": "array"
          },
          "window": {
            "$ref": "#/components/schemas/Window"
          }
        },
        "required": [
          "meta",
          "window",
          "months",
          "totals",
          "top_categories",
          "expense_trend",
          "weekly",
          "daily"
        ],
        "title": "Trends",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Window": {
        "properties": {
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "label": {
            "description": "e.g. \"FY 2026-27 to date\", \"Apr 2025 – 25 Sep 2026\"",
            "title": "Label",
            "type": "string"
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end",
          "label"
        ],
        "title": "Window",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Clario API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/connections/{connection_id}/finance/balance-sheet": {
      "get": {
        "operationId": "get_balance_sheet_api_v1_workspaces__workspace_id__connections__connection_id__finance_balance_sheet_get",
        "parameters": [
          {
            "description": "Connection id",
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "description": "Connection id",
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceSheet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Balance Sheet tab (pending the director's definition)",
        "tags": [
          "finance"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/connections/{connection_id}/finance/gst": {
      "get": {
        "operationId": "get_gst_api_v1_workspaces__workspace_id__connections__connection_id__finance_gst_get",
        "parameters": [
          {
            "description": "Connection id",
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "description": "Connection id",
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GstOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "GST tab (pending the director's definition)",
        "tags": [
          "finance"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/connections/{connection_id}/finance/invoices": {
      "get": {
        "operationId": "get_invoices_api_v1_workspaces__workspace_id__connections__connection_id__finance_invoices_get",
        "parameters": [
          {
            "description": "Connection id",
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "description": "Connection id",
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "overdue",
                    "unpaid",
                    "paid",
                    "partially_paid",
                    "open",
                    "draft",
                    "void"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "party",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Party"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 500,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoicePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invoice register, newest first",
        "tags": [
          "finance"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/connections/{connection_id}/finance/overview": {
      "get": {
        "operationId": "get_overview_api_v1_workspaces__workspace_id__connections__connection_id__finance_overview_get",
        "parameters": [
          {
            "description": "Connection id",
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "description": "Connection id",
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overview tab",
        "tags": [
          "finance"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/connections/{connection_id}/finance/receivables": {
      "get": {
        "operationId": "get_receivables_api_v1_workspaces__workspace_id__connections__connection_id__finance_receivables_get",
        "parameters": [
          {
            "description": "Connection id",
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "description": "Connection id",
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receivables"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Receivables tab",
        "tags": [
          "finance"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/connections/{connection_id}/finance/trends": {
      "get": {
        "operationId": "get_trends_api_v1_workspaces__workspace_id__connections__connection_id__finance_trends_get",
        "parameters": [
          {
            "description": "Connection id",
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "description": "Connection id",
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trends"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trends & Analysis tab",
        "tags": [
          "finance"
        ]
      }
    },
...
  }
}